+++ catalog/app/components/FileEditor/index.ts
@@ -1,3 +1,4 @@
 export * from './Controls'
 export * from './CreateFile'
 export * from './FileEditor'
+export * from './State'
